svelte-migrate: updated files
