+++ src/routes/+error.svelte
@@ -1,28 +1,19 @@
-<script context="module">
-	export function load({ error, status }) {
-		return {
-			props: { error, status }
-		};
-	}
-</script>
-
 <script>
 	import { dev } from '$app/env';
-
-	export let error, status;
+	import { page } from '$app/stores';
 </script>
 
 <svelte:head>
-	<title>{status}</title>
+	<title>{$page.status}</title>
 </svelte:head>
 
 <div class="col-md-9">
-	<h1>{status}</h1>
+	<h1>{$page.status}</h1>
 
-	<p>{error.message}</p>
+	<p>{$page.error.message}</p>
 
-	{#if dev && error.stack}
-		<pre>{error.stack}</pre>
+	{#if dev && $page.error.stack}
+		<pre>{$page.error.stack}</pre>
 	{/if}
 </div>
