+++ src/routes/artifacts/[id]/+page.svelte
@@ -1,4 +1,4 @@
-<script context="module" lang="ts">
+<script lang="ts">
 	import { page } from '$app/stores';
 	import artifactDetails from '$lib/graphql/customStores/artifactDetails';
 	import {
@@ -7,19 +7,11 @@
 		KQL_ArtifactPatch
 	} from '$lib/graphql/_kitql/graphqlStores';
 	import { somethingWentWrong } from '$lib/utils';
-	import { KitQLInfo } from '@kitql/all-in';
 	import { onMount } from 'svelte';
 	import { writable } from 'svelte/store';
-	import DetailBase from './_DetailBase.svelte';
-	import { emptyArtifact } from './_utils';
+	import DetailBase from '../_DetailBase.svelte';
+	import { emptyArtifact } from '../_utils';
 
-	export const load = async ({ fetch, params }) => {
-		await artifactDetails.queryLoad({ fetch, variables: { id: params.id } });
-		return {};
-	};
-</script>
-
-<script lang="ts">
 	const { id } = $page.params;
 	const variables = { id }; // for requests
