+++ src/routes/characters/+page.js
@@ -1,0 +1,6 @@
+import { KQL_Npcs } from '$lib/graphql/_kitql/graphqlStores';
+
+export const load = async ({ fetch }) => {
+	await KQL_Npcs.queryLoad({ fetch });
+	return {};
+};
