+++ src/routes/characters/+page.svelte
@@ -1,17 +1,10 @@
-<script context="module" lang="ts">
+<script>
 	import AddLink from '$lib/components/AddLink.svelte';
 	import BannerImage from '$lib/components/BannerImage.svelte';
 	import ListDetailCard from '$lib/components/ListDetailCard.svelte';
 	import { KQL_Npcs } from '$lib/graphql/_kitql/graphqlStores';
 	import { alphabetically } from '$lib/utils';
 
-	export const load = async ({ fetch }) => {
-		await KQL_Npcs.queryLoad({ fetch });
-		return {};
-	};
-</script>
-
-<script>
 	$: npcs = $KQL_Npcs.data?.npcs.edges?.map(({ node }) => node).sort(alphabetically) || [];
 </script>
