+++ src/routes/endpoints/apirel/+server.ts
@@ -1,4 +1,4 @@
-import api from './_api.js';
+import api from '../_api.js';
 
 /**
  * Endpoint to access the graphql backend
