+++ src/routes/places/+page.js
@@ -1,0 +1,6 @@
+import { KQL_Places } from '$lib/graphql/_kitql/graphqlStores';
+
+export const load = async ({ fetch }) => {
+	await KQL_Places.queryLoad({ fetch });
+	return {};
+};
