+++ src/routes/races/+page.js
@@ -1,0 +1,6 @@
+import { KQL_Races } from '$lib/graphql/_kitql/graphqlStores';
+
+export const load = async ({ fetch }) => {
+	await KQL_Races.queryLoad({ fetch });
+	return {};
+};
