+++ src/routes/races/+page.svelte
@@ -1,4 +1,4 @@
-<script context="module" lang="ts">
+<script>
 	import AddLink from '$lib/components/AddLink.svelte';
 	import BannerImage from '$lib/components/BannerImage.svelte';
 	import ListDetailCard from '$lib/components/ListDetailCard.svelte';
@@ -6,13 +6,6 @@
 	import { alphabetically } from '$lib/utils';
 	import { compass } from '@cloudinary/url-gen/qualifiers/gravity';
 
-	export const load = async ({ fetch }) => {
-		await KQL_Races.queryLoad({ fetch });
-		return {};
-	};
-</script>
-
-<script>
 	$: races = $KQL_Races.data?.races.edges?.map(({ node }) => node).sort(alphabetically) || [];
 </script>
