+++ src/routes/races/[id]/+page.js
@@ -1,0 +1,6 @@
+import raceDetails from '$lib/graphql/customStores/raceDetails';
+
+export const load = async ({ fetch, params }) => {
+	await raceDetails.queryLoad({ fetch, variables: { id: params.id } });
+	return {};
+};
