+++ src/routes/races/[id]/+page.svelte
@@ -1,21 +1,13 @@
-<script context="module" lang="ts">
+<script>
 	import { page } from '$app/stores';
 	import raceDetails from '$lib/graphql/customStores/raceDetails';
 	import { KQL_RaceAddImage, KQL_RaceLock, KQL_RacePatch } from '$lib/graphql/_kitql/graphqlStores';
 	import { somethingWentWrong } from '$lib/utils';
-	import { KitQLInfo } from '@kitql/all-in';
 	import { onMount } from 'svelte';
 	import { writable } from 'svelte/store';
-	import DetailBase from './_DetailBase.svelte';
-	import { emptyRace } from './_utils';
+	import DetailBase from '../_DetailBase.svelte';
+	import { emptyRace } from '../_utils';
 
-	export const load = async ({ fetch, params }) => {
-		await raceDetails.queryLoad({ fetch, variables: { id: params.id } });
-		return {};
-	};
-</script>
-
-<script>
 	const { id } = $page.params;
 	const variables = { id }; // for requests
